encryption decryption enhancement

+++ src/main/java/dataman/dmbase/dto/FileDTO.java
@@ -14,6 +14,6 @@ public class FileDTO{
 	
 	private byte[] data;
 	private String fileType;
-	private Long fileSize;
+	private double fileSize;
 	
 }
